requirements.txt
fastapi
uvicorn
pydantic
python-multipart
streamlit
Pillow
numpy
scikit-learn
pandas
joblib
matplotlib
locust
